Elasticsearch: Rename public/app/plugins/datasource/elasticsearch/language_provider.ts to follow convention (#55430)
